Feature/1829/smoke test directories (#1287)
Smoke tests split up by directory and updated the nightly job #4306
