+++ cypress/integration/smokeTests/sharedTests/search.ts
